fix : ReadGroup의 EditMode와 ViewMode를 분리
