+++ app/src/main/java/kr/nbc/momo/presentation/group/read/EditReadGroupViewModel.kt
@@ -8,16 +8,11 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.launch
-import kr.nbc.momo.domain.usecase.JoinGroupUseCase
-import kr.nbc.momo.domain.usecase.BlockUserUseCase
 import kr.nbc.momo.domain.usecase.ChangeLeaderUseCase
 import kr.nbc.momo.domain.usecase.DeleteGroupUseCase
 import kr.nbc.momo.domain.usecase.DeleteUserUseCase
 import kr.nbc.momo.domain.usecase.ReadGroupUseCase
-import kr.nbc.momo.domain.usecase.ReportUserUseCase
 import kr.nbc.momo.domain.usecase.UpdateGroupUseCase
-import kr.nbc.momo.domain.usecase.AddUserUseCase
-import kr.nbc.momo.domain.usecase.SubscriptionUseCase
 import kr.nbc.momo.presentation.UiState
 import kr.nbc.momo.presentation.group.mapper.asGroupEntity
 import kr.nbc.momo.presentation.group.mapper.toGroupModel
@@ -25,15 +20,18 @@ import kr.nbc.momo.presentation.group.model.GroupModel
 import javax.inject.Inject
 
 @HiltViewModel
-class ReadGroupViewModel @Inject constructor(
+class EditReadGroupViewModel @Inject constructor(
     private val readGroupUseCase: ReadGroupUseCase,
-    private val subscriptionUseCase: SubscriptionUseCase,
+    private val updateGroupUseCase: UpdateGroupUseCase,
     private val deleteGroupUseCase: DeleteGroupUseCase,
-    private val reportUserUseCase: ReportUserUseCase,
-    private val blockUserUseCase: BlockUserUseCase,
+    private val changeLeaderUseCase: ChangeLeaderUseCase,
+    private val deleteUserUseCase: DeleteUserUseCase
 ) : ViewModel() {
     private val _groupState = MutableStateFlow<UiState<GroupModel>>(UiState.Loading)
     val groupState: StateFlow<UiState<GroupModel>> get() = _groupState
+
+    private val _updateState = MutableStateFlow<UiState<GroupModel>>(UiState.Loading)
+    val updateState: StateFlow<UiState<GroupModel>> get() = _updateState
 
     private val _subscriptionState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
     val subscriptionState: StateFlow<UiState<Boolean>> get() = _subscriptionState
@@ -41,11 +39,12 @@ class ReadGroupViewModel @Inject constructor(
     private val _deleteGroupState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
     val deleteGroupState: StateFlow<UiState<Boolean>> get() = _deleteGroupState
 
-    private val _reportUserState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
-    val reportUserState: StateFlow<UiState<Boolean>> get() = _reportUserState
 
-    private val _blockUserState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
-    val blockUserState: StateFlow<UiState<Boolean>> get() = _blockUserState
+    private val _leaderChangeState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
+    val leaderChangeState: StateFlow<UiState<Boolean>> get() = _leaderChangeState
+
+    private val _userDeleteState = MutableStateFlow<UiState<List<String>>>(UiState.Loading)
+    val userDeleteState: StateFlow<UiState<List<String>>> get() = _userDeleteState
 
     fun readGroup(groupId: String) {
         viewModelScope.launch {
@@ -66,20 +65,20 @@ class ReadGroupViewModel @Inject constructor(
         }
     }
 
-    fun subscription(userId: String, groupId: String) {
-        viewModelScope.launch {
-            _subscriptionState.value = UiState.Loading
 
-            subscriptionUseCase.invoke(userId, groupId)
+    fun updateGroup(groupModel: GroupModel, imageUri: Uri?) {
+        viewModelScope.launch {
+            _updateState.value = UiState.Loading
+
+            updateGroupUseCase.invoke(groupModel.asGroupEntity(), imageUri)
                 .catch { e ->
-                    _subscriptionState.value = UiState.Error(e.toString())
+                    _updateState.value = UiState.Error(e.toString())
                 }
                 .collect { data ->
-                    _subscriptionState.value = UiState.Success(data)
+                    _updateState.value = UiState.Success(data.toGroupModel())
                 }
         }
     }
-
 
     fun deleteGroup(groupId: String, userList: List<String>) {
         viewModelScope.launch {
@@ -95,33 +94,32 @@ class ReadGroupViewModel @Inject constructor(
         }
     }
 
-    fun reportUser(reportedUser: String) {
-        viewModelScope.launch {
-            _reportUserState.value = UiState.Loading
 
-            reportUserUseCase.invoke(reportedUser)
+    fun leaderChange(groupId: String, leaderId: String) {
+        viewModelScope.launch {
+            _leaderChangeState.value = UiState.Loading
+
+            changeLeaderUseCase.invoke(groupId, leaderId)
                 .catch { e ->
-                    _reportUserState.value = UiState.Error(e.toString())
+                    _leaderChangeState.value = UiState.Error(e.toString())
                 }
                 .collect { data ->
-                    _reportUserState.value = UiState.Success(data)
+                    _leaderChangeState.value = UiState.Success(data)
                 }
         }
     }
 
-    fun blockUser(blockUser: String) {
+    fun deleteUser(userId: String, groupId: String) {
         viewModelScope.launch {
-            _blockUserState.value = UiState.Loading
+            _userDeleteState.value = UiState.Loading
 
-            blockUserUseCase.invoke(blockUser)
+            deleteUserUseCase.invoke(userId, groupId)
                 .catch { e ->
-                    _blockUserState.value = UiState.Error(e.toString())
+                    _userDeleteState.value = UiState.Error(e.toString())
                 }
                 .collect { data ->
-                    _blockUserState.value = UiState.Success(data)
+                    _userDeleteState.value = UiState.Success(data)
                 }
         }
     }
-
-
 }
