+++ app/src/main/java/kr/nbc/momo/presentation/home/HomeFragment.kt
@@ -263,7 +263,7 @@ class HomeFragment : Fragment() {
                 val readGroupFragment = ReadGroupFragment()
                 parentFragmentManager.beginTransaction()
                     .replace(R.id.fragment_container, readGroupFragment)
-                    .addToBackStack(null)
+                    .addToBackStack("Read")
                     .commit()
             }
         }
@@ -275,7 +275,7 @@ class HomeFragment : Fragment() {
                 val readGroupFragment = ReadGroupFragment()
                 parentFragmentManager.beginTransaction()
                     .replace(R.id.fragment_container, readGroupFragment)
-                    .addToBackStack(null)
+                    .addToBackStack("Read")
                     .commit()
             }
         }
@@ -287,7 +287,7 @@ class HomeFragment : Fragment() {
                 val readGroupFragment = ReadGroupFragment()
                 parentFragmentManager.beginTransaction()
                     .replace(R.id.fragment_container, readGroupFragment)
-                    .addToBackStack(null)
+                    .addToBackStack("Read")
                     .commit()
             }
         }
